Rename helpers to utils/make isNew a private attribute
